refactor and remove JS dependence
