+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableWebChromeClient.java
@@ -1,0 +1,17 @@
+package com.iterable.iterableapi;
+
+import android.webkit.WebChromeClient;
+import android.webkit.WebView;
+
+public class IterableWebChromeClient extends WebChromeClient {
+    IterableWebView.HTMLNotificationCallbacks inAppHTMLNotification;
+
+    IterableWebChromeClient(IterableWebView.HTMLNotificationCallbacks inAppHTMLNotification) {
+        this.inAppHTMLNotification = inAppHTMLNotification;
+    }
+
+    @Override
+    public void onProgressChanged(WebView view, int newProgress) {
+        inAppHTMLNotification.recalculateHeight(view.getContentHeight());
+    }
+}
